# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PuntosFieles API",
    "version": "1.3.7",
    "description": "Route-derived OpenAPI document for versioned /api/v1 routes, with curated request/response metadata layered onto the discovered route surface."
  },
  "servers": [
    {
      "url": "https://api.puntos.local"
    }
  ],
  "paths": {
...
    "/api/v1/admin/branches": {
      "get": {
        "summary": "List branches for the current business",
        "responses": {
          "200": {
            "description": "Branch list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Plan feature not enabled or role not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "multi_branch"
        ]
      },
      "post": {
        "summary": "Create a branch",
        "responses": {
          "200": {
            "description": "Branch created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Plan feature not enabled or role not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-csrf-required": true,
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "multi_branch"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BranchCreateRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/external-awards": {
      "get": {
        "summary": "Get external award integration settings",
        "responses": {
          "200": {
            "description": "External awards configuration"
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Owner role and plan feature required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Business not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "external_awards"
        ]
      },
      "put": {
        "summary": "Update external award integration settings",
        "responses": {
          "200": {
            "description": "External awards updated"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Owner role and plan feature required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Business not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-csrf-required": true,
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "external_awards"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalAwardsRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/signup": {
      "post": {
        "summary": "Derived POST /api/v1/admin/signup",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ]
      }
    },
...
    "/api/v1/admin/webhooks": {
      "get": {
        "summary": "List webhook endpoints for the current business",
        "responses": {
          "200": {
            "description": "Webhook endpoint list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "webhooks"
        ]
      },
      "post": {
        "summary": "Create a webhook endpoint",
        "responses": {
          "200": {
            "description": "Webhook endpoint created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Role not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "staffAuth": []
          }
        ],
        "x-csrf-required": true,
        "x-tenant-context": true,
        "x-required-roles": [
          "OWNER"
        ],
        "x-plan-features": [
          "webhooks"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreateRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "staffAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pf_staff"
      },
      "customerAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pf_customer"
      },
      "superAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pf_super"
      }
    },
    "schemas": {
...
      "BranchCreateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "address": {
            "type": "string",
            "maxLength": 250
          },
          "code": {
            "type": "string",
            "minLength": 3,
            "maxLength": 80
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
...
      "WebhookCreateRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "secret": {
            "type": "string",
            "minLength": 8
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 1
          }
        },
        "required": [
          "url",
          "secret",
          "events"
        ],
        "additionalProperties": false
      },
...
      "ExternalAwardsRequest": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "api_key": {
            "type": "string",
            "minLength": 8,
            "maxLength": 120
          }
        },
        "required": [
          "enabled"
        ],
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "BranchListResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "branches": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "business_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "name": {
                  "type": "string"
                },
                "address": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "code": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "code"
              ],
              "additionalProperties": true
            }
          }
        },
        "required": [
          "ok",
          "branches"
        ],
        "additionalProperties": false
      },
      "BranchDetailResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "branch": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "business_id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              },
              "address": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "code": {
                "type": "string"
              },
              "created_at": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name",
              "code"
            ],
            "additionalProperties": true
          }
        },
        "required": [
          "ok",
          "branch"
        ],
        "additionalProperties": false
      },
...
      "WebhookListResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "endpoints": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "business_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "url": {
                  "type": "string",
                  "format": "uri"
                },
                "events": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "active": {
                  "type": "boolean"
                },
                "created_at": {
                  "type": "string"
                },
                "secret_masked": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "has_secret": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "url",
                "events",
                "active"
              ],
              "additionalProperties": true
            }
          }
        },
        "required": [
          "ok",
          "endpoints"
        ],
        "additionalProperties": false
      },
      "WebhookDetailResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "endpoint": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "business_id": {
                "type": "string",
                "format": "uuid"
              },
              "url": {
                "type": "string",
                "format": "uri"
              },
              "events": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "active": {
                "type": "boolean"
              },
              "created_at": {
                "type": "string"
              },
              "secret_masked": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "has_secret": {
                "type": "boolean"
              }
            },
            "required": [
              "id",
              "url",
              "events",
              "active"
            ],
            "additionalProperties": true
          }
        },
        "required": [
          "ok",
          "endpoint"
        ],
        "additionalProperties": false
      },
...
    }
  }
}